refactor(native): extract sidecar cache lifecycle
